alarm: Implemented an alarm to schedule and execute tasks periodically
At this moment, after the alarm is turned on, it will display a message
at every 10 seconds.

The setAlarm() method implemented in AlarmService class starts the alarm
and the cancelAlarm() method stops the alarm.

If the alarm is set and after a while the application is being closed,
the alarm will still be active and when the application is reopened again,
the alarm can be deactivate by calling the cancelAlarm() method.

If the phone is rebooted and before the reboot the alarm was active,
the alarm will not be active after the reboot.

Signed-off-by: Codrut Grosu <codrut.cristian.grosu@gmail.com>